# avx2: Add 64 bytes per loop processing